fixup: reorg waypoint utility into motion_planning module
